Added All Tasks
Added the code to accomplish the asks for all tasks that require coding. Tasks 3 and 4 only require terminal commands to be run and their output observed. All C files were compiled with gcc with no special arguments. Only the standard libraries are used. For the bruteforce python script, pycrypto is needed for all the AES operations.

The keys file for the bruteforce python script is not included and can be generated with task2.c because where's the fun in giving out all the answers.